+++ src/main/java/com/sapo/mockprojectpossystem/service/interfaces/IProductImageService.java
@@ -7,11 +7,11 @@ import java.io.IOException;
 import java.util.List;
 
 public interface IProductImageService {
-    ProductImageResponse createProductImage(Long productId, MultipartFile image) throws IOException;
+    ProductImageResponse createProductImage(Integer productId, MultipartFile image) throws IOException;
 
-    ProductImageResponse getProductImageById(Long id, Long productId);
+    ProductImageResponse getProductImageById(Integer id, Integer productId);
 
-    List<ProductImageResponse> getAllImagesWithProductId(Long productId);
+    List<ProductImageResponse> getAllImagesWithProductId(Integer productId);
 
-    void deleteProductImageById(Long id, Long productId) throws Exception;
+    void deleteProductImageById(Integer id, Integer productId) throws Exception;
 }
