+++ src/main/java/com/sapo/mockprojectpossystem/service/interfaces/IProductVariantService.java
@@ -6,11 +6,11 @@ import com.sapo.mockprojectpossystem.dto.response.ProductVariantResponse;
 import java.util.List;
 
 public interface IProductVariantService {
-    List<ProductVariantResponse> getAllProductVariantsByProductId(Long productId);
+    List<ProductVariantResponse> getAllProductVariantsByProductId(Integer productId);
 
-    ProductVariantResponse getProductVariantById(Long id, Long productId);
+    ProductVariantResponse getProductVariantById(Integer id, Integer productId);
 
-    void deleteProductVariantById(Long id, Long productId);
+    void deleteProductVariantById(Integer id, Integer productId);
 
-    ProductVariantResponse updateProductVariant(Long id, Long productId, ProductVariantRequest request);
+    ProductVariantResponse updateProductVariant(Integer id, Integer productId, ProductVariantRequest request);
 }
